chore: extract pure player component so i can move other things around
