+++ frontend/src/scenes/session-recordings/player/SessionRecordingPlayer.scss
@@ -1,6 +1,46 @@
+.SessionRecordingPlayerWrapper {
+    position: relative;
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+
+    .SessionRecordingPlayer__sidebar {
+        position: relative;
+        display: flex;
+        flex-direction: column;
+        flex-shrink: 0;
+        min-width: 20rem;
+        max-width: 95%;
+        overflow: hidden;
+
+        &--collapsed {
+            min-width: 0;
+        }
+
+        &__tabs {
+            & .LemonTabs__tab-content {
+                padding: 10px 0;
+            }
+        }
+    }
+
+    &--stacked-vertically {
+        flex-direction: column;
+
+        .SessionRecordingPlayer__sidebar {
+            min-width: 100%;
+            max-width: 100%;
+            max-height: 60%;
+        }
+    }
+}
+
 .SessionRecordingPlayer {
     position: relative;
     display: flex;
+    flex: 1;
     flex-direction: row;
     width: 100%;
     height: 100%;
@@ -24,7 +64,10 @@
     }
 
     .SessionRecordingPlayer__main {
+        display: flex;
         flex: 1;
+        flex-direction: column;
+        overflow: hidden;
     }
 
     &--fullscreen {
@@ -43,43 +86,9 @@
         }
     }
 
-    .SessionRecordingPlayer__main,
-    .SessionRecordingPlayer__sidebar {
-        display: flex;
-        flex-direction: column;
-        overflow: hidden;
-    }
-
     .LemonModal & {
         .SessionRecordingPlayer__body {
             height: calc(90vh - 15rem);
-        }
-    }
-
-    .SessionRecordingPlayer__sidebar {
-        position: relative;
-        flex-shrink: 0;
-        min-width: 20rem;
-        max-width: 95%;
-
-        &--collapsed {
-            min-width: 0;
-        }
-
-        &__tabs {
-            & .LemonTabs__tab-content {
-                padding: 10px 0;
-            }
-        }
-    }
-
-    &--stacked-vertically {
-        flex-direction: column;
-
-        .SessionRecordingPlayer__sidebar {
-            min-width: 100%;
-            max-width: 100%;
-            max-height: 60%;
         }
     }
 }
